Rename shellout module
