+++ mocilib/shellouts.py
